Moves `Use template` to the top header
* Creates a separate component for header component
